requirements.txt
django==4.1.1
djangorestframework==3.13.1
Pillow==9.2.0
psycopg2-binary==2.9.3
python-dotenv==0.21.0
pandas==1.4.4
numpy==1.23.2
matplotlib==3.5.3
django-cors-headers==3.13.0
